+++ NodeCityList.java
@@ -13,7 +13,7 @@ public class NodeCityList extends LinkedList<String>{
     }
 
     //Add to list of flights
-    public int addHead(String s, int c, int m){
+    public int addHead(String s, float c, int m){
         Node<String> e = new Node<>(s,c,m);
         e.next = this.head;
         this.head = e;
